+++ tsconfig.json
@@ -19,6 +19,7 @@
 		"./typings/index.d.ts",
 		"./src/actions/storeTodoActions.ts",
 		"./src/actions/uiTodoActions.ts",
+		"./src/actions/widgetTodoActions.ts",
 		"./src/index.ts",
 		"./src/main.ts",
 		"./src/registry/createTodoRegistry.ts",
